moved gif encoder implementation to lib/src/ for dartdevc optimisation
